Adds framework for syncing to the config and main loop, reorganizes the endpoints
