+++ google_classroom/endpoints/__init__.py
@@ -1,0 +1,31 @@
+from endpoints.announcement import Announcement
+from endpoints.course import Course
+from endpoints.course_alias import CourseAlias
+from endpoints.coursework import CourseWork
+from endpoints.guardian import Guardian
+from endpoints.guardian_invitation import GuardianInvitation
+from endpoints.invitation import Invitation
+from endpoints.meet import Meet
+from endpoints.org_unit import OrgUnit
+from endpoints.student import Student
+from endpoints.submission import StudentSubmission
+from endpoints.student_usage import StudentUsage
+from endpoints.teacher import Teacher
+from endpoints.topic import Topic
+
+__all__ = [
+    "Announcement",
+    "Course",
+    "CourseAlias",
+    "CourseWork",
+    "Guardian",
+    "GuardianInvitation",
+    "Invitation",
+    "Meet",
+    "OrgUnit",
+    "Student",
+    "StudentSubmission",
+    "StudentUsage",
+    "Teacher",
+    "Topic",
+]
