+++ google_classroom/endpoints/course.py
@@ -1,0 +1,33 @@
+from endpoints.base import EndPoint
+
+
+class Course(EndPoint):
+    def __init__(self, service, sql, config):
+        super().__init__(service, sql, config)
+        self.date_columns = ["creationTime", "updateTime"]
+        self.columns = [
+            "id",
+            "name",
+            "courseGroupEmail",
+            "courseState",
+            "creationTime",
+            "description",
+            "descriptionHeading",
+            "enrollmentCode",
+            "guardiansEnabled",
+            "ownerId",
+            "room",
+            "section",
+            "teacherGroupEmail",
+            "updateTime",
+        ]
+        self.request_key = "courses"
+        self.batch_size = config.COURSES_BATCH_SIZE
+
+    def request_data(self, course_id=None, date=None, next_page_token=None):
+        return self.service.courses().list(
+            pageToken=next_page_token, pageSize=self.config.PAGE_SIZE,
+        )
+
+    def filter_data(self, dataframe):
+        return dataframe[dataframe.updateTime >= self.config.SCHOOL_YEAR_START]
