+++ google_classroom/endpoints/guardian.py
@@ -1,0 +1,18 @@
+from endpoints.base import EndPoint
+
+
+class Guardian(EndPoint):
+    def __init__(self, service, sql, config):
+        super().__init__(service, sql, config)
+        self.columns = ["studentId", "guardianId", "invitedEmailAddress"]
+        self.request_key = "guardians"
+        self.batch_size = config.GUARDIANS_BATCH_SIZE
+
+    def request_data(self, course_id=None, date=None, next_page_token=None):
+        return (
+            self.service.userProfiles()
+            .guardians()
+            .list(
+                studentId="-", pageToken=next_page_token, pageSize=self.config.PAGE_SIZE
+            )
+        )
